Add mval and data kwargs to TelemData and add data kwarg to Node
The new mval kwarg is a boolean that determines if the TelemData
component supplies a row to mvals.  If telemdata is used in the core
model then mvals must be True, otherwise it can be False.  The
advantage to mvals=False is that the telemetry data need not be
read until it gets used instead of when model.make() is called.

The new data kwarg allows for setting the initial data value for
a TelemData or Node component.  This facilitates non-telemetry
values because the initial value gets stored in the model spec.