Refactor (mostly) to make sherpa fitting work within PyQt4
I ran into a strange problem when doing sherpa fitting of a
xija model in the same process where a Qt event loop was
running.  Basically if I did a ui.fit(1) OR
(set the user stat model and evaluated the xija model) then
this would kill the main window for no obvious reason.

So I refactored to get the model definition and everything
else needed for fitting into the process prior to making a
multiprocess Pool, but BEFORE starting the Qt event loop.
The idea was to make a clean process for model fitting that
doesn't have Qt going.  This seems to work at least prior to
trying to get the interprocess communication via pipes working
again...